pyproject.toml
[tool.poetry]
name = "vroozi-purchase-order-processing"
version = "0.1.0"
description = "Automated purchase order processing system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = "^0.21.1"
motor = "^3.1.1"
redis = "^4.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-multipart = "^0.0.6"
pydantic = {extras = ["email", "dotenv"], version = "^1.10.7"}
prometheus-fastapi-instrumentator = "^5.9.0"
slowapi = "^0.1.5"
structlog = "^23.1.0"
email-validator = "^2.0.0"
mongoengine = "0.27.0"
pymongo = "4.5.0"
circuitbreaker = "^1.4.0"
fastapi-limiter = "^0.1.5"
backoff = "^2.2.1"
google-cloud-vision = "^3.4.0"
boto3 = "^1.26.0"
jinja2 = "^3.1.2"
weasyprint = "^57.1"
python-docx = "^0.8.11"
aiosmtplib = "^2.0.0"
celery = "^5.2.7"
sentry-sdk = "^1.14.0"
psutil = "^5.9.0"
pdf2image = "^1.17.0"
tenacity = "^8.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py39"]
include = "\\.pyi?$"

[tool.isort]
profile = "black"
multi_line_output = 3